cart refacto part 2
